requirement.txt
Flask==2.3.2
requests==2.31.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0